% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{adapt_round}
\alias{adapt_round}
\title{Adaptive rounding for tables
This is an unvectorized form that works and gives single value (as opposed to the original function).}
\usage{
adapt_round(x)
}
\arguments{
\item{x}{a numeric constant}
}
\value{
an adaptively rounded constant.
}
\description{
Adaptive rounding for tables
This is an unvectorized form that works and gives single value (as opposed to the original function).
}
\examples{
adapt_round(0.3454)
}
