% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\docType{data}
\name{amy_pal}
\alias{amy_pal}
\title{Amyloidosis paper 4-5-color scheme}
\format{
An object of class \code{character} of length 6.
}
\usage{
amy_pal
}
\description{
Amyloidosis paper 4-5-color scheme
}
\keyword{datasets}
