% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{formatp}
\alias{formatp}
\title{Format p-values conventionally
uses rndformat() from this pckg <0.001, <0.01.}
\usage{
formatp(p)
}
\arguments{
\item{p}{}
}
\value{
pretty print character vector, rounded p
}
\description{
Format p-values conventionally
uses rndformat() from this pckg <0.001, <0.01.
}
