% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{formatp_nejm}
\alias{formatp_nejm}
\title{Format p-values per NEJM style
uses rndformat() from this pckg <0.0001, <0.001}
\usage{
formatp_nejm(p)
}
\arguments{
\item{p}{}
}
\value{
pretty print character vector, rounded p
}
\description{
Format p-values per NEJM style
uses rndformat() from this pckg <0.0001, <0.001
}
