% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{geom_split_violin}
\alias{geom_split_violin}
\title{Split violin plot
a violin plot with violin split in two for two groups}
\usage{
geom_split_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "identity",
  ...,
  draw_quantiles = NULL,
  trim = TRUE,
  scale = "area",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{inherit.aes}{}
}
\value{
a violin plot with violin split in two for two groups
}
\description{
Split violin plot
a violin plot with violin split in two for two groups
}
