% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\docType{data}
\name{GeomSplitViolin}
\alias{GeomSplitViolin}
\title{ggproto object needed for split violin graph}
\format{
An object of class \code{GeomSplitViolin} (inherits from \code{GeomViolin}, \code{Geom}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
GeomSplitViolin
}
\description{
ggproto object needed for split violin graph
}
\keyword{datasets}
