% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{get_axislabel}
\alias{get_axislabel}
\title{Get proper axis label for outcome variable
Named vector as look-up table}
\usage{
get_axislabel(var, lookupvector = names_long)
}
\arguments{
\item{var}{"fida", "fidv", "fica", "avr", "crae", "crve", "fmd", "gtn", "pwv" or "aix"}

\item{lookupvector}{names_long is saved in fun}
}
\value{
axis label for plot
}
\description{
Get proper axis label for outcome variable
Named vector as look-up table
}
