% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{ggplotl}
\alias{ggplotl}
\title{ggplotl
Wrapper around ggplot() to use labels as axis labels}
\usage{
ggplotl(...)
}
\arguments{
\item{...}{ggplot parameters as usual}
}
\value{
ggplot with attr-label as axis labels
}
\description{
ggplotl
Wrapper around ggplot() to use labels as axis labels
}
