% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{ihs}
\alias{ihs}
\title{Inverse hyperbolic sine transformation
Can normalize data by log(x + sqrt(x ^ 2 + 1)}
\usage{
ihs(x)
}
\arguments{
\item{x}{variable (e. g. vector) to be transformed}
}
\value{

}
\description{
Inverse hyperbolic sine transformation
Can normalize data by log(x + sqrt(x ^ 2 + 1)
}
