% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{impmed}
\alias{impmed}
\title{Simple median imputation
Multiple imputation with Amelia & mice is better for a set with several missing data points.}
\usage{
impmed(x)
}
\arguments{
\item{x}{variable with missing values}
}
\value{
imputed data with median of existing data plus random number
}
\description{
Simple median imputation
Multiple imputation with Amelia & mice is better for a set with several missing data points.
}
