% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{my_html}
\alias{my_html}
\title{html-ify sMobj by F. Harrell
e.g. summaryM object}
\usage{
my_html(sMobj, caption)
}
\arguments{
\item{caption}{}
}
\value{

}
\description{
html-ify sMobj by F. Harrell
e.g. summaryM object
}
