% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{mykablestyle}
\alias{mykablestyle}
\title{kable_styling wrapper to ensure all tables are consistently styled
e.g. summaryM object from Hmisc}
\usage{
mykablestyle(obj, stripes = FALSE, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
kable_styling wrapper to ensure all tables are consistently styled
e.g. summaryM object from Hmisc
}
