% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{retinal}
\alias{retinal}
\title{retinal palette colors ramped to a specified length}
\usage{
retinal(palette = "aurora", n, alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'jens_palettes' list}

\item{n}{Number of colors to display}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
retinal palette colors ramped to a specified length
}
\examples{
library(jenshelper)
image(volcano, col = retinal("aurora", 20))
}
