% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{retinal_pal}
\alias{retinal_pal}
\title{Retinal palette with ramped colours}
\usage{
retinal_pal(palette = "aurora", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from 'retinal_palettes' list}

\item{alpha}{transparency}

\item{reverse}{If TRUE, the direction of the colours is reversed.}
}
\description{
Retinal palette with ramped colours
}
