% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{rms_results_kable}
\alias{rms_results_kable}
\title{RMS result table
by Frank Harrell}
\usage{
rms_results_kable(rmsObj, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
RMS result table
by Frank Harrell
}
