% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{rndformat}
\alias{rndformat}
\title{Round format}
\usage{
rndformat(x, digits = 2, ...)
}
\arguments{
\item{x}{a numeric}

\item{digits}{to be rounded to}

\item{...}{handed over to base fun format}
}
\value{
quoted "pretty print" character vector
}
\description{
Round format
}
