% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{se}
\alias{se}
\title{Standard error of the mean}
\usage{
se(x, sd = NULL, na.rm = TRUE)
}
\arguments{
\item{sd}{true population sd if known, else sample sd autom. calculated}

\item{na.rm}{default is TRUE!}
}
\value{
the standard error (sd/root n), not the Student's approx.!
}
\description{
Standard error of the mean
}
