% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{tableone_mat_to_data_frame}
\alias{tableone_mat_to_data_frame}
\title{T1 mat to dataframe
Turn tableone output matrix into tidyverse data_frame}
\usage{
tableone_mat_to_data_frame(mat)
}
\arguments{
\item{mat}{}
}
\value{

}
\description{
T1 mat to dataframe
Turn tableone output matrix into tidyverse data_frame
}
