% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{theme_jens}
\alias{theme_jens}
\title{My ggplot2 Theme}
\usage{
theme_jens(
  font_size = 20,
  font_family = "Humanist 521",
  line_size = 1,
  rel_small = 14/16,
  rel_tiny = 12/14,
  rel_large = 16/14
)
}
\arguments{
\item{font_size}{in pt}

\item{font_family}{font or either of sans, serif}

\item{line_size}{width of important lines}

\item{rel_small}{legend, strip and subtitle}

\item{rel_tiny}{for captions}

\item{rel_large}{for title}

\item{portrait}{If F sets margins adequate for landscape plots (assymmetric)}
}
\value{
a graph theme to be added to a ggplot
}
\description{
It's based on cowplot which modifies theme_grey. Doesn't depend on cowplot. It %+replace% in theme_grey.
Particular: y-axis label is title. ylab is supposed to be the unit of abscissa.
}
