% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{using}
\alias{using}
\title{Using
Loads all the packages, then goes back and installs all the missing packages.
Does not make use of installed.packages as suggested in R documentation (slow!).
I run it before I install pacman.
Code from: https://stackoverflow.com/questions/4090169/elegant-way-to-check-for-missing-packages-and-install-them}
\usage{
using(...)
}
\arguments{
\item{...}{package (CRAN) in quotes}
}
\value{
attach or install & attach library
}
\description{
Using
Loads all the packages, then goes back and installs all the missing packages.
Does not make use of installed.packages as suggested in R documentation (slow!).
I run it before I install pacman.
Code from: https://stackoverflow.com/questions/4090169/elegant-way-to-check-for-missing-packages-and-install-them
}
