% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{write_df_to_xlsx}
\alias{write_df_to_xlsx}
\title{Write a data frame to an xlsx file (TableOne)}
\usage{
write_df_to_xlsx(mat, file, font_size = 9)
}
\arguments{
\item{mat}{a matrix, i. e. from TableOne}

\item{file}{destination file}

\item{font_size}{}
}
\value{
an Excel file with Table
#'
}
\description{
Write a data frame to an xlsx file (TableOne)
}
