% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{write_tableone_to_xlsx}
\alias{write_tableone_to_xlsx}
\title{Write a matrix via data frame to an xlsx file
For Tableone objects: It writes a tableone (w/o print) to Excel file}
\usage{
write_tableone_to_xlsx(mat, file)
}
\arguments{
\item{font_size}{}
}
\value{
an .xlsx file
}
\description{
Write a matrix via data frame to an xlsx file
For Tableone objects: It writes a tableone (w/o print) to Excel file
}
