% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jenshelper.R
\name{jb_figure}
\alias{jb_figure}
\title{Plot survey bar plots}
\usage{
jb_figure(df = des_amy, out = fida, upper = NA, pal = "tol3")
}
\arguments{
\item{df}{a survey design obsject (survey or srvyr)}

\item{out}{a single variable of interest}

\item{upper}{custom max value of y axis}

\item{fillncolor}{any palette from \code{\link[jenshelper]{retinal_palettes}}}
}
\value{
a ggplot dynamite plot with errorbars
}
\description{
Plot survey bar plots
}
